Test - redux toolkit 적용 및 slice.test 방법 수정
slice.test를 reducer에 대한 테스트가 아니라 store 자체에 대한 테스트로
수정
