+++ src/LoginFormContainer.test.jsx
@@ -14,46 +14,42 @@ describe('LoginFormContainer', () => {
 
     useDispatch.mockImplementation(() => dispatch);
 
-    useSelector.mockImplementation((selector) => selector({
-      loginFields: {
-        email: 'test@test',
-        password: '1234',
-      },
-      accessToken: given.accessToken,
-    }));
+    useSelector.mockImplementation((selector) =>
+      selector({
+        loginFields: {
+          email: 'test@test',
+          password: '1234',
+        },
+        accessToken: given.accessToken,
+      })
+    );
   });
 
   context('when logged out', () => {
     given('accessToken', () => '');
 
     it('renders input controls', () => {
-      const { getByLabelText } = render((
-        <LoginFormContainer />
-      ));
+      const { getByLabelText } = render(<LoginFormContainer />);
 
       expect(getByLabelText('E-mail').value).toBe('test@test');
       expect(getByLabelText('Password').value).toBe('1234');
     });
 
     it('listens change events', () => {
-      const { getByLabelText } = render((
-        <LoginFormContainer />
-      ));
+      const { getByLabelText } = render(<LoginFormContainer />);
 
       fireEvent.change(getByLabelText('E-mail'), {
         target: { value: 'new email' },
       });
 
       expect(dispatch).toBeCalledWith({
-        type: 'changeLoginField',
+        type: 'app/changeLoginField',
         payload: { name: 'email', value: 'new email' },
       });
     });
 
     it('renders “Log In” button', () => {
-      const { getByText } = render((
-        <LoginFormContainer />
-      ));
+      const { getByText } = render(<LoginFormContainer />);
 
       fireEvent.click(getByText('Log In'));
 
@@ -65,13 +61,11 @@ describe('LoginFormContainer', () => {
     given('accessToken', () => 'ACCESS_TOKEN');
 
     it('renders “Log out” button', () => {
-      const { getByText } = render((
-        <LoginFormContainer />
-      ));
+      const { getByText } = render(<LoginFormContainer />);
 
       fireEvent.click(getByText('Log out'));
 
-      expect(dispatch).toBeCalledWith({ type: 'logout' });
+      expect(dispatch).toBeCalledWith({ type: 'app/logout' });
     });
   });
 });
