+++ src/RestaurantsPage.jsx
@@ -8,9 +8,7 @@ import RegionsContainer from './RegionsContainer';
 import CategoriesContainer from './CategoriesContainer';
 import RestaurantsContainer from './RestaurantsContainer';
 
-import {
-  loadInitialData,
-} from './actions';
+import { loadInitialData } from './slice';
 
 export default function RestaurantsPage() {
   const history = useHistory();
